Splitting rendering concern
